stop using dep
It is currently not possible to have API interfaces in the "vendor" folder. If two or more projects share the same interfaces via their own vendor folders, the go compiler will not treat the interfaces as equal. The reason is the different paths to the sources on the filesystem. There is an open issue on Github here https://github.com/golang/go/issues/18827.